restful_onlinetvshop/requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coverage==4.5.3
Django==2.2.1
django-crispy-forms==1.7.2
django-fontawesome==1.0
django-modelcluster==4.4
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.9.4
draftjs-exporter==2.1.5
html5lib==1.0.1
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
psycopg2==2.8.2
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
stripe==2.28.0
typed-ast==1.3.5
urllib3==1.24.3
webencodings==0.5.1
wrapt==1.11.1
